refactor: Change Activity percentage fields to int
This commit changes the data types of `progress_planned_percent` and
`progress_achieved_percent` fields in the `Activity` model from `float`
to `int` (with a range of 0-100) throughout the project. This aligns
them with similar percentage fields like KR metas.

Changes include:
- Updated `Activity` Pydantic Model (`app/models.py`):
    - `progress_planned_percent` and `progress_achieved_percent` are now `int`.
    - Default value for `progress_achieved_percent` updated to `0`.
- Verified Activity Service (`app/services/activity_service.py`):
    - No code changes were necessary as existing f-string formatting for
      these fields in `_serialize_activity_to_table_row` correctly
      handles the new integer types from the model.
- Updated Unit Tests (`tests/unit/test_activity_service.py`):
    - Test data and payloads for `Activity` instances now use integer values
      for these percentages.
    - String assertions for Markdown table rows updated to expect integer
      representations (e.g., "100%" instead of "100.0%").
- Updated Integration Tests (`tests/integration/test_api.py`):
    - Payloads in `test_03_add_activities_to_kr` now use integers for
      these activity percentage fields.
    - Assertions on KR description strings (containing the activity table)
      were reviewed to ensure they align with integer formatting.
- Updated API Documentation (`docs/api_requirements_diagram.md`):
    - The description of the `Activity` model in Section 4 now explicitly
      lists `progress_planned_percent` and `progress_achieved_percent`
      as `int`.
- Verified Postman Collection (`docs/postman_collection.json`):
    - The example payload for "Add Activities to KR" was confirmed to
      already use integer values for these fields from a previous update.

+++ app/models.py
@@ -49,8 +49,8 @@ class Activity(BaseModel):
     stakeholders: str
     deadline_planned: str # e.g., "Month/Year" or specific date string
     deadline_achieved: Optional[str] = None
-    progress_planned_percent: float = Field(..., ge=0, le=100)
-    progress_achieved_percent: float = Field(default=0.0, ge=0, le=100)
+    progress_planned_percent: int = Field(..., ge=0, le=100)
+    progress_achieved_percent: int = Field(default=0, ge=0, le=100)
 
 class ActivityCreateRequest(BaseModel):
     activities: List[Activity]
